Feat: Implementado outras exceções

+++ src/main/java/com/example/fabio/dscommerce/services/ProductService.java
@@ -3,11 +3,15 @@ package com.example.fabio.dscommerce.services;
 import com.example.fabio.dscommerce.dto.ProductDTO;
 import com.example.fabio.dscommerce.entities.Product;
 import com.example.fabio.dscommerce.repositories.ProductRepository;
+import com.example.fabio.dscommerce.services.exceptions.DatabaseException;
 import com.example.fabio.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -46,19 +50,28 @@ public class ProductService {
 
     @Transactional()
     public ProductDTO update(Long id, ProductDTO dto) {
+        try {
 
         Product entity = repository.getReferenceById(id);
         copyDtoToEntity(dto, entity);
-
         entity = repository.save(entity);
-
         return new ProductDTO(entity);
+
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
-    @Transactional()
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
-
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
